requirements.txt
openai==1.56.1
loguru
tiktoken
GitPython
tree-sitter==0.20.4
rich
html2text
fire
pre-commit
